backend/requirements.txt
Flask
flask-mysqldb
flask-cors
flask-jwt-extended
python-dotenv
Werkzeug
bleach
gunicorn
